SamuelSanchezArroyoPortfolio/ClientApp/src/app/counter/counter.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { CounterComponent } from './counter.component';

describe('CounterComponent', () => {
  let fixture: ComponentFixture<CounterComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ CounterComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CounterComponent);
    fixture.detectChanges();
  });

  it('should display a title', async(() => {
    const titleText = fixture.nativeElement.querySelector('h1').textContent;
    expect(titleText).toEqual('Counter');
  }));

  it('should start with count 0, then increments by 1 when clicked', async(() => {
    const countElement = fixture.nativeElement.querySelector('strong');
    expect(countElement.textContent).toEqual('0');

    const incrementButton = fixture.nativeElement.querySelector('button');
    incrementButton.click();
    fixture.detectChanges();
    expect(countElement.textContent).toEqual('1');
  }));
});

describe('CounterComponent2', () => {
  let fixture2: ComponentFixture<CounterComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [CounterComponent]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture2 = TestBed.createComponent(CounterComponent);
    fixture2.detectChanges();
  });

  it('should display a title', async(() => {
    const titleText = fixture2.nativeElement.querySelector('h1').textContent;
    expect(titleText).toEqual('Counter');
  }));

  it('should start with count 0, then increments by 2 when clicked', async(() => {
    const countElement = fixture2.nativeElement.querySelector('strong');
    expect(countElement.textContent).toEqual('0');

    const incrementButton2 = fixture2.nativeElement.querySelector('button');
    incrementButton2.click();
    fixture2.detectChanges();
    expect(countElement.textContent).toEqual('0');
  }));
});
